backend/requirements.txt
easyocr
pillow
numpy
fastapi
uvicorn
pydantic
requests
openai
dotenv
google-cloud-logging
